Add support for local versioning schemes
PEP-440 allows for a local portion of a version string that
follows the public version. This version is set off from the
public version with a '+' and can consist of arbitrarily many
'.' delimited segments. Each segment must consist solely of
ASCII characters [a-zA-Z0-9]. Segments consisting solely of
digits should be compared numerically while those containin
letters should be compared lexically. Numeric segments should
always be considered "greater than" lexical segments. All
prefix components being equal, the string with more segments
should be considered greater.

This patch adds support for that feature, along with tests
that ensure this feature remains in compliance with the rest
of the infrastructure.